+++ client/api/MovieApi.js
@@ -4,13 +4,13 @@ const apiClient = axios.create({
     baseURL:"http://localhost:8000"
 });
 
-export function getTopMovies() {
+export function getTopMoviesApi() {
     let promise = apiClient.get('/getTopMovies/');
     const dataPromise = promise.then((response) => response.data);
     return dataPromise;
 }
 
-export function searchMovies(query) {
+export function searchMoviesApi(query) {
     // headers = {'Content-type': 'application/json', 'Accept': 'application/json'}
     let data = {"query":query};
     console.log("data", data);
